Fix out-of-bounds panic.
Series that don't have any chunks anymore after flushing are invalid and
need to be removed (in the Prometheus storage, it's an invariant that a
series always has at least one chunk loaded into memory or the series is
evicted completely). Otherwise, this crashed when calling head() on the
series.