Add POC Distributor->Ingestor HTTP query API.
I want to attempt to add an official /api/v1/query_raw endpoint to
Prometheus, so for now this is lifting code from the Prometheus API and
clientlib in an improvised way for now:

- vendored HTTP API client library was changed in place to allow
  raw queries
- private helper types from Prometheus's HTTP API were copied over for
  implementing the raw query endpoint.

Obviously, this can only be temporary. The first wart would go away if
we get the raw query interface into Prometheus and then modify the
actual clientlib to look like the modified vendored version here. For
the second point, we'll have to restructure the API-serving code in
Prometheus in such a way that it's possible to only reuse the raw query
part of it.