Add metric for number of ingester memory chunks
I considered exporting the number of samples stored in the chunk store
as well, but the problem is that not all chunk types support directly
getting their sample count anymore, so that method was removed from the
chunk interface. Thus getting the number of samples from a chunk upon
storing it would require iterating over the entire chunk, which doesn't
seem worth it. But at least with the number of memory chunks, we should
also notice backlogging behavior.